+++ dora/services/management/commands/send_saved_searchs_notifications.py
@@ -1,4 +1,4 @@
-from datetime import datetime, timedelta
+from datetime import timedelta
 
 from django.conf import settings
 from django.core.management.base import BaseCommand
@@ -8,8 +8,8 @@ from mjml import mjml2html
 
 from dora import data_inclusion
 from dora.core.emails import send_mail
-from dora.services.models import SavedSearch, SavedSearchFrequency
-from dora.services.views import _search
+
+from ...models import SavedSearch, SavedSearchFrequency
 
 
 def get_saved_search_notifications_to_send():
@@ -67,50 +67,18 @@ class Command(BaseCommand):
 
         num_emails_sent = 0
         for saved_search in saved_searchs:
-            category = None
-            if saved_search.category:
-                category = saved_search.category
-
-            subcategories = None
-            if saved_search.subcategories.exists():
-                subcategories = saved_search.subcategories.values_list(
-                    "value", flat=True
-                )
-
-            kinds = None
-            if saved_search.kinds.exists():
-                kinds = saved_search.kinds.values_list("value", flat=True)
-
-            fees = None
-            if saved_search.fees.exists():
-                fees = saved_search.fees.values_list("value", flat=True)
-
-            # Récupération des résultats de la recherche
-            results = _search(
-                None,
-                saved_search.city_code,
-                [category.value] if category and not subcategories else None,
-                subcategories,
-                kinds,
-                fees,
-                di_client,
+            # On garde les contenus qui ont été publiés depuis la dernière notification
+            new_services = saved_search.get_recent_services(
+                saved_search.last_notification_date, di_client
             )
 
-            # On garde les contenus qui ont été publiés depuis la dernière notification
-            updated_services = [
-                r
-                for r in results
-                if datetime.fromisoformat(r["publication_date"]).date()
-                > saved_search.last_notification_date
-            ]
-
-            if updated_services:
+            if new_services:
                 # Envoi de l'email
                 context = {
                     "search_label": compute_search_label(saved_search),
                     "homepage_url": settings.FRONTEND_URL,
-                    "updated_services": updated_services[:15],
-                    "services_number": len(updated_services),
+                    "updated_services": new_services[:15],
+                    "services_number": len(new_services),
                 }
 
                 num_emails_sent += 1
